make new test_distributed.py
